+++ trikot-viewmodels-declarative-flow/compose-flow/src/main/kotlin/com/mirego/trikot/viewmodels/declarative/compose/viewmodel/VMDImage.kt
@@ -37,15 +37,15 @@ fun VMDImage(
     placeholderContentScale: ContentScale = contentScale,
     alpha: Float = DefaultAlpha,
     colorFilter: ColorFilter? = null,
-    contentDescription: String? = null,
     allowHardware: Boolean = true,
     placeholder: @Composable ((placeholderImageResource: VMDImageResource, state: AsyncImagePainter.State) -> Unit) = { imageResource, state ->
+        val imageViewModel by viewModel.observeAsState(excludedProperties = if (modifier.isOverridingAlpha()) listOf(viewModel::isHidden) else emptyList())
         RemoteImageDefaultPlaceholder(
             imageResource = imageResource,
             modifier = modifier,
             contentScale = placeholderContentScale,
             colorFilter = colorFilter,
-            contentDescription = contentDescription
+            contentDescription = imageViewModel.contentDescription
         )
     },
     asyncStateCallback: ((AsyncImagePainter.State) -> Unit)? = null
@@ -60,7 +60,7 @@ fun VMDImage(
         alignment = alignment,
         contentScale = contentScale,
         colorFilter = colorFilter,
-        contentDescription = contentDescription,
+        contentDescription = imageViewModel.contentDescription,
         placeholderContentScale = placeholderContentScale,
         imageDescriptor = imageViewModel.image,
         allowHardware = allowHardware,