+++ trikot-viewmodels-declarative-flow/viewmodels-declarative-flow/src/commonMain/kotlin/com/mirego/trikot/viewmodels/declarative/components/impl/VMDImageViewModelImpl.kt
@@ -17,13 +17,22 @@ open class VMDImageViewModelImpl(coroutineScope: CoroutineScope) :
         emit(VMDImageDescriptor.Local(VMDImageResource.None) as VMDImageDescriptor, this, coroutineScope)
     override var image: VMDImageDescriptor by imageDelegate
 
+    private val contentDescriptionDelegate: VMDFlowProperty<String?> =
+        emit(null, this, coroutineScope)
+    override var contentDescription: String? by contentDescriptionDelegate
+
     fun bindImage(flow: Flow<VMDImageDescriptor>) {
         updateProperty(this::image, flow)
+    }
+
+    fun bindContentDescription(flow: Flow<String?>) {
+        updateProperty(this::contentDescription, flow)
     }
 
     override val propertyMapping: Map<String, VMDFlowProperty<*>> by lazy {
         super.propertyMapping.toMutableMap().also {
             it[this::image.name] = imageDelegate
+            it[this::contentDescription.name] = contentDescriptionDelegate
         }
     }
 }