+++ trikot-viewmodels-declarative/viewmodels-declarative/src/commonMain/kotlin/com/mirego/trikot/viewmodels/declarative/components/impl/VMDImageViewModelImpl.kt
@@ -17,13 +17,22 @@ open class VMDImageViewModelImpl(cancellableManager: CancellableManager) :
         published(VMDImageDescriptor.Local(VMDImageResource.None) as VMDImageDescriptor, this)
     override var image: VMDImageDescriptor by imageDelegate
 
+    private val contentDescriptionDelegate: VMDPublishedProperty<String?> =
+        published(null, this)
+    override var contentDescription: String? by contentDescriptionDelegate
+
     fun bindImage(publisher: Publisher<VMDImageDescriptor>) {
         updatePropertyPublisher(this::image, cancellableManager, publisher)
+    }
+
+    fun bindContentDescription(publisher: Publisher<String?>) {
+        updatePropertyPublisher(this::contentDescription, cancellableManager, publisher)
     }
 
     override val propertyMapping: Map<String, VMDPublishedProperty<*>> by lazy {
         super.propertyMapping.toMutableMap().also {
             it[this::image.name] = imageDelegate
+            it[this::contentDescription.name] = contentDescriptionDelegate
         }
     }
 }